[CKV2_AWS_76] Ensure AWS ALB attached WAFv2 WebACL is configured with AMR for Log4j Vulnerability
Introduce a new "environment" variable in the WAF module to improve configurability. Update the WAF module to include environment tagging and connect its ACL to the ELB module in the "develop" environment setup. Also, document a Checkov skip rule in the ELB module for Log4J protection managed by WAF.